feat: restore yandere command, add filters to yandere command
